save-file.use-case test

+++ 04-multiplicationApp/src/domain/use-cases/save-file.use-case.test.ts
@@ -3,8 +3,20 @@ import fs from 'fs';
 
 describe('save-file.use-case', () => {
   
+  const customOptions = {
+    fileContent: 'custom content',
+    fileDestination: 'custom-outputs/file-destination',
+    fileName: 'custom-table-name'
+  }
+
+  const customFilePath = `${customOptions.fileDestination}/${customOptions.fileName}.txt`;
+
   afterEach(() => {
-    fs.rmSync('outputs', { recursive: true })
+    const outputFolderExists = fs.existsSync('outputs'); 
+    if( outputFolderExists) fs.rmSync('outputs', { recursive: true });
+
+    const customOutputFolderExists = fs.existsSync(customOptions.fileDestination);
+    if(customOutputFolderExists) fs.rmSync(customOptions.fileDestination, {recursive: true})
   });
 
   test('should save file with default values', () => { 
@@ -22,4 +34,18 @@ describe('save-file.use-case', () => {
     expect(fileExists).toBe(true);
     expect(fileContent).toBe(options.fileContent)
   })
+
+  test('should save file with custom values', () => {
+
+    const saveFile = new SaveFile();
+    
+    const result = saveFile.execute(customOptions);
+    const fileExists = fs.existsSync(customFilePath);
+    const fileContent = fs.readFileSync(customFilePath, {encoding: 'utf-8'})
+
+    expect(result).toBe(true);
+    expect(fileExists).toBe(true);
+    expect(fileContent).toBe(customOptions.fileContent)
+  })
+
 });
